Set reply-to queue address before post-processor
Because setting the reply-to address in the post-processor forces the
users of custom post-processors to remember/add a line to set the
reply-to address always. By setting the reply-to prior to the
post-processor, the user has the option to modify the reply-to before
sending, if desired. By setting the reply-to prior to post-processor, we
free up some cognitive load on the users who wish to use a custom
post-processor.