Fix race condition in AmqpConnection
The automatic reconnection  was racy. It could try to open a connection
at the same time that `Close()` was called on `AmqpConnection`. This
required introducing an internal field to determine when the connection
is "closed", and a specific error to signal in the reconnection that the
connection is "closed" after calling `Close()`.