RPC high level constructs (#59)
* Set Go toolchain

* Initial version of RPC server

The happy path works. The server can connect to a queue and responds
messages if the request's replyTo property is set.

* Bump Ginkgo Gomega and use floating 4.1 image in test

* Add VS Code gitignores

* Add correlationID extractor and post-processor

RPC Tests were refactored to extract common setup code into a before
each node.

* RPC server test for custom functionality

A test that verifies that custom correction ID and custom post processor
are called by the server.

* Add code docs for RPC server functions

The test for RpcServer.Close() was returning too soon. The simplest
solution is an ugly sleep of 1 second.

* Add functions to handle consumer credits manually

RPC server construct provides an utility to pause the message
processing. This relies on the ability to "drain" credits, and issue new
credits at a later point to unpause. This commit exposes that
functionality internally.

* Implement pause/unpause in RPC server

The RPC server now can be constructred from the AMQP connection. Tests
were refactored to use this contructor, instead of manually setting up
publishers and consumers.

Because setting the reply-to address in the post-processor forces the
users of custom post-processors to remember/add a line to set the
reply-to address always. By setting the reply-to prior to the
post-processor, the user has the option to modify the reply-to before
sending, if desired. By setting the reply-to prior to post-processor, we
free up some cognitive load on the users who wish to use a custom
post-processor.

* Add E2E test for RPC constructs

* Refactor pointer-to functions: one generic function to rule them all
* The automatic reconnection  was racy. It could try to open a connection
at the same time that `Close()` was called on `AmqpConnection`. This
required introducing an internal field to determine when the connection
is "closed", and a specific error to signal in the reconnection that the
connection is "closed" after calling `Close()`.
* Add examples and fix typo
Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>

---------

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>
Co-authored-by: Gabriele Santomaggio <G.santomaggio@gmail.com>