Refactor and test fixes (#66)
* Refactor and remove QueueType

It is just a wrapper around TQueueType

* Fix test assertions running outside subject nodes

Assertions should be made only inside subject nodes i.e. It(), otherwise, a failed assertion panics and the test code does not recover correctly.

* Refactor exchange type

It had a wrapper struct. It's simpler to use the exchange type directly

* Fix assertions running outside subject node

Because it can cause panics on failed assertions, and it does not respect the focus marker i.e. it always used to run.