Mark topology E2E test as flaky
Because the management helper sometimes fails to find the connection string.
This seems to be alleviated by increasing the timeouts to 10 seconds,
and polling to 500ms. This makes sense because the management API in the
browser refreshes every 5 seconds, and rabbit probably needs some time
to update the management DB with connection information.

With this commit, the E2E test suite ran successfully 26 times in a row.